exceptions: Rename py files task_allocation -> allocation, task_execution -> execution
