+++ mrs/allocation/round.py
@@ -4,8 +4,8 @@ import time
 from datetime import timedelta
 
 import numpy as np
-from mrs.exceptions.task_allocation import AlternativeTimeSlot
-from mrs.exceptions.task_allocation import NoAllocation
+from mrs.exceptions.allocation import AlternativeTimeSlot
+from mrs.exceptions.allocation import NoAllocation
 from mrs.bidding.bid import Bid
 from ropod.utils.timestamp import TimeStamp
 from ropod.utils.uuid import generate_uuid
